Feature update
(Filter by Platform): New dropdown with streaming platform options. clicking on one filters results to specified platform.
Checkbox in dashboard allows for certain recommendations to be filtered out based on the platform their available on the specified platform
Backend performs a search using the tmdbsimple watchprovider endpoint or the direct url (tmdb simple wrapper was giving me issues with tv shows)
The search gives information about which provider the given media is available on. A Boolean is returned based on the input from the front end.
Backend has a map of provider ids that are used for the conversion from input names to provider ids.
When the filter is used, results are saved for 10 minutes. query_string is enabled so different combinations are cached differently
For example if Netflix and hulu are selected, it will use the is_available_on_platform function to find which of the recommendations are valid.
The second time its asked for that combination, the result will appear instantly.
When using the platform filter, user recs are limited to 20 to prevent API limits. (time.sleep(0.1) is a secondary precaution along the primary safeguard of the flask limiter)

Platform selections are saved in sessionStorage and serve as the default once set.

Loading spinner added to search and dashboard to account for longer loading time when using platform filter

*Next addition is a button to view liked movies or followed tv*

(Backend Optimizations)
Using Flask Blueprints to separate files
Working with one files with everything in it was getting difficult.
Teleshow.py is still the main file. It will automatically use the blueprints when needed.
Sentence transformer is loaded when needed to decrease startup latency
Firebase db is loaded up when needed to decrease startup latency
Recommendations algorithm is optimized to use less API calls and use a less computationally taxing learning model.
(Using a static model that faster and a lot more efficient)

One .env file for both frontend and backend.

*Filter in dashboard is planned to be moved to a settings panel
*View of liked media for editing is planned for the settings panel